pyproject.toml
[project]
name = "httpingest"
version = "53.0.0"
description = "LogicMonitor OTLP data pipeline for ingesting metrics to Azure Data Lake Gen2"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-identity>=1.19.0",
    "azure-storage-file-datalake>=12.14.0",
    "pyarrow>=14.0.0",
    "fastapi>=0.121.2",
    "uvicorn[standard]>=0.32.0",
    # OpenTelemetry tracing
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
    "opentelemetry-instrumentation-logging>=0.41b0",
]

[dependency-groups]
dev = [
    "aiohttp>=3.13.2",
    "duckdb>=1.0.0",
    "httpx>=0.27.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "module"
asyncio_default_test_loop_scope = "module"
